Refactor: libpacemaker: best practices for applying orderings
stage7() in pcmk_sched_allocate.c previously had two unrelated components,
applying orderings and logging actions.

The ordering component is now pcmk__apply_orderings() in pcmk_sched_ordering.c
(declared in the library-private), and the logging component is now
log_all_actions() in pcmk_sched_messages.c. Comments and formatting have been
improved.

This abandons the "stage" naming scheme, since it was already broken (e.g. a
stage 0 and no stage 1, with extraneous stuff inserted into various "stages").
