+++ include/pcmki/pcmki_scheduler.h
@@ -72,7 +72,6 @@ extern gboolean stage3(pe_working_set_t * data_set);
 extern gboolean stage4(pe_working_set_t * data_set);
 extern gboolean stage5(pe_working_set_t * data_set);
 extern gboolean stage6(pe_working_set_t * data_set);
-extern gboolean stage7(pe_working_set_t * data_set);
 extern gboolean stage8(pe_working_set_t * data_set);
 
 extern gboolean unpack_constraints(xmlNode * xml_constraints, pe_working_set_t * data_set);
