+++ lib/pacemaker/libpacemaker_private.h
@@ -92,6 +92,9 @@ void pcmk__order_stops_before_shutdown(pe_node_t *node,
                                        pe_action_t *shutdown_op,
                                        pe_working_set_t *data_set);
 
+G_GNUC_INTERNAL
+void pcmk__apply_orderings(pe_working_set_t *data_set);
+
 /*!
  * \internal
  * \brief Create a new ordering between two resource actions
@@ -120,5 +123,8 @@ void pcmk__order_stops_before_shutdown(pe_node_t *node,
 
 G_GNUC_INTERNAL
 void pcmk__unpack_rsc_ticket(xmlNode *xml_obj, pe_working_set_t *data_set);
+
+G_GNUC_INTERNAL
+void pcmk__order_probes(pe_working_set_t *data_set);
 
 #endif // PCMK__LIBPACEMAKER_PRIVATE__H
