+++ lib/pacemaker/pcmk_sched_messages.c
@@ -21,6 +21,7 @@
 #include <crm/pengine/status.h>
 #include <pacemaker-internal.h>
 #include <crm/common/ipc_internal.h>
+#include "libpacemaker_private.h"
 
 extern bool pcmk__is_daemon;
 
@@ -48,6 +49,34 @@ log_resource_details(pe_working_set_t *data_set)
     }
 
     g_list_free(all);
+}
+
+static void
+log_all_actions(pe_working_set_t *data_set)
+{
+    /* This only ever outputs to the log, so ignore whatever output object was
+     * previously set and just log instead.
+     */
+    pcmk__output_t *prev_out = data_set->priv;
+    pcmk__output_t *out = pcmk__new_logger();
+
+    if (out == NULL) {
+        return;
+    }
+
+    pcmk__output_set_log_level(out, LOG_NOTICE);
+    data_set->priv = out;
+
+    out->begin_list(out, NULL, NULL, "Actions");
+    LogNodeActions(data_set);
+
+    g_list_foreach(data_set->resources, (GFunc) LogActions, data_set);
+
+    out->end_list(out);
+    out->finish(out, CRM_EX_OK, true, NULL);
+    pcmk__output_free(out);
+
+    data_set->priv = prev_out;
 }
 
 /*!
@@ -105,7 +134,8 @@ pcmk__schedule_actions(pe_working_set_t *data_set, xmlNode *xml_input,
     crm_trace("Processing fencing and shutdown cases");
     stage6(data_set);
 
-    stage7(data_set); // Ordering constraints
+    pcmk__apply_orderings(data_set);
+    log_all_actions(data_set);
 
     crm_trace("Create transition graph");
     stage8(data_set);
